+++ runtime/service/src/main/java/org/apache/polaris/service/config/DefaultConfigurationStore.java
@@ -25,15 +25,13 @@ import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import java.util.Map;
 import java.util.Optional;
-import org.apache.polaris.core.config.PolarisConfigurationStore;
 import org.apache.polaris.core.config.RealmConfigurationSource;
 import org.apache.polaris.core.context.RealmContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @ApplicationScoped
-public class DefaultConfigurationStore
-    implements PolarisConfigurationStore, RealmConfigurationSource {
+public class DefaultConfigurationStore implements RealmConfigurationSource {
   Logger LOGGER = LoggerFactory.getLogger(DefaultConfigurationStore.class);
 
   private final Map<String, Object> defaults;
@@ -54,13 +52,6 @@ public class DefaultConfigurationStore
     LOGGER.debug("Get configuration value for {} with realm {}", configName, realm);
     return Optional.ofNullable(realmOverrides.getOrDefault(realm, Map.of()).get(configName))
         .orElseGet(() -> getDefaultConfiguration(configName));
-  }
-
-  @Override
-  public <T> @Nullable T getConfiguration(@Nonnull RealmContext realmContext, String configName) {
-    @SuppressWarnings("unchecked")
-    T value = (T) getConfigValue(realmContext, configName);
-    return value;
   }
 
   private <T> @Nullable T getDefaultConfiguration(String configName) {